mqtt-broker/bridge/requirements.txt
paho-mqtt==2.1.0
confluent-kafka==2.4.0
python-dotenv==1.0.1
pyyaml==6.0.2
